refactor: rename `dashboard-layout.tsx` into `DashboardLayout.tsx`
