+++ app/(main)/pilkada2024/page.tsx
@@ -1,4 +1,4 @@
-import DashboardLayout from '@/components/dashboard-layout'
+import DashboardLayout from '@/components/DashboardLayout'
 import MapDashboardProvider from '@/modules/MapDashboard/DashboardProvider'
 import MapView from '@/modules/Pilkada2024/MapView'
 import PilkadaProvider from '@/modules/Pilkada2024/PilkadaProvider'
